Add py.typed file to package
Add a py.typed file to the package denoting that typing is supported per
PEP 561.

According to mypy documentation, zip_safe=False must be passed to
setuptools in setup(), or else mypy will not be able to find the
installed package.

This was tested by building a source dist via:

    python setup.py sdist

And then installing the generated tarball and running mypy on a dummy
Python file.